dom/indexedDB/test/test_indexes.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      const CONSTRAINT_ERR =
        Components.interfaces.nsIIDBDatabaseException.CONSTRAINT_ERR;
      const READ_WRITE = Components.interfaces.nsIIDBTransaction.READ_WRITE;

      const name = window.location.pathname;
      const description = "My Test Database";

      const objectStoreName = "People";

      const objectStoreData = [
        { key: "237-23-7732", value: { name: "Bob", height: 60, weight: 120 } },
        { key: "237-23-7733", value: { name: "Ann", height: 52, weight: 110 } },
        { key: "237-23-7734", value: { name: "Ron", height: 73, weight: 180 } },
        { key: "237-23-7735", value: { name: "Sue", height: 58, weight: 130 } },
        { key: "237-23-7736", value: { name: "Joe", height: 65, weight: 150 } },
        { key: "237-23-7737", value: { name: "Pat", height: 65 } }
      ];

      const indexData = [
        { name: "name", keyPath: "name", unique: true },
        { name: "height", keyPath: "height", unique: false },
        { name: "weight", keyPath: "weight", unique: false }
      ];

      let request = indexedDB.open(name, description);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let db = event.result;

      request = db.createObjectStore(objectStoreName, "");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      let objectStore = event.result;

      // First, add all our data to the object store.
      let addedData = 0;
      for (let i in objectStoreData) {
        request = objectStore.add(objectStoreData[i].value,
                                  objectStoreData[i].key);
        request.onerror = errorHandler;
        request.onsuccess = function(event) {
          if (++addedData == objectStoreData.length) {
            testGenerator.send(event);
          }
        }
      }
      event = yield;

      // Now create the index.
      addedData = 0;
      for (let i in indexData) {
        request = objectStore.createIndex(indexData[i].name,
                                          indexData[i].keyPath,
                                          indexData[i].unique);
        request.onerror = errorHandler;
        request.onsuccess = function(event) {
          if (++addedData == indexData.length) {
            is(objectStore.indexNames.length, addedData, "Good index count");
            SimpleTest.executeSoon(function() { testGenerator.next() });
          }
        }
      }
      yield;

      objectStore = db.objectStore(objectStoreName);

      // Check global properties to make sure they are correct.
      is(objectStore.indexNames.length, indexData.length, "Good index count");
      for (let i in indexData) {
        let found = false;
        for (let j = 0; j < objectStore.indexNames.length; j++) {
          if (objectStore.indexNames.item(j) == indexData[i].name) {
            found = true;
            break;
          }
        }
        is(found, true, "objectStore has our index");
        let index = objectStore.index(indexData[i].name);
        is(index.name, indexData[i].name, "Correct name");
        is(index.storeName, objectStore.name, "Correct store name");
        is(index.keyPath, indexData[i].keyPath, "Correct keyPath");
        is(index.unique, indexData[i].unique, "Correct keyPath");
      }

      request = objectStore.index("name").get("Bob");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result, "237-23-7732", "Correct key returned!");

      request = objectStore.index("name").getObject("Bob");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result.name, "Bob", "Correct name returned!");
      is(event.result.height, 60, "Correct height returned!");
      is(event.result.weight, 120, "Correct weight returned!");

      objectStore = db.objectStore(objectStoreName, READ_WRITE);
      request = objectStore.add({ name: "Bob", height: 62, weight: 170 },
                                "237-23-7738");
      request.onerror = new ExpectError(CONSTRAINT_ERR);
      request.onsuccess = unexpectedSuccessHandler;
      event = yield;

      finishTest();
      yield;
    }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>
</head>

<body onload="runTest();"></body>

</html>
